argument encoder refactored, tests moved to verification.test.ts and test runner can now run any .test.ts file in test folder
